restructured working tree
